Move getter and setter for the state of completedness for the importers in the abstract importing action
